Omorganiserer litt
